bugfix,: added auth confirmation for deleting budget

+++ backend/internal/services/budget_service.go
@@ -111,7 +111,14 @@ func (bs *BudgetServices) BudgetList(user_id uuid.UUID) ([]models.Budget, error)
 }
 
 // BudgetDelete removes a budget by ID using the repository.
-func (bs *BudgetServices) BudgetDelete(budget_id uuid.UUID) error {
+func (bs *BudgetServices) BudgetDelete(budget_id, user_id uuid.UUID) error {
+	budget, err := bs.budgetRepo.GetBudgetByID(budget_id)
+	if err != nil {
+		return  err
+	}
+	if budget.UserID != user_id {
+		return utils.ErrForbidden
+	} 
 	if err := bs.budgetRepo.DeleteBudget(budget_id); err != nil {
 		return err
 	}
